kernel: add support for the lz4hc compressor
This is a compressor in the Linux kernel which provides benefits similar
to lz4 (the fastest compressor available in the kernel), but with a better
compression ratio.

This compressor shares code with lz4, they both use the same
decompressor, therefore, there is not a big impact in image size. This
package depends on lz4 and will select it.

By enabling this compressor, users can select this method and benefit
from it's performance when using zram or zswap, given that it's
objectively the second fastest compressor/decompressor available.

Signed-off-by: Oever González <notengobattery@gmail.com>