compressed-memory: add packages for compressed memory
This commit adds support for a new compressed memory management for
OpenWrt. Thep purpose of this package is to replace zram-swap with the
scripts provided in here. However, it's not a direct replacement, it
will allow users to use zram-swap, or replace it with this.

In this package, the following installable packages are included:
 - swap-zram-writeback: script for managing zram with writeback support
 - swap-zswap-cache: script for managing zswap compressed cache

Both packages are independent and can be used alone, but they also can
be used torgether and provide advanced interoperability when both
alternatives are used simultaneusly.

This package requieres a series of commits to enable the requiered
kernel featueres, therefore, this change cannot be applied alone. Still,
this package adds dependency to losetup for swap-zram-writeback. Both of
the new packages will conflict with zram-swap.

The advantages of this package over zram-swap:
 - Support for zram writeback (backing device) to reduce memory usage
   when configured correctly. Disabled by default, requieres ROM
   storage for the backing file.
 - Advanced and extensible UCI interface. Adding more compressors and
   managing the pool is trivial an it's more resistant to errors and
   misconfigurations.
 - Features can be enabled or disabled from the UCI interface, providing
   better LuCI integration, without the need to disable the startup
   scripts.
 - Support for zswap. This can be useful combined with zram, or with a
   real swap device. The configuration is easier than zram, but
   requieres a real swapping device (it can be the swap from zram, a
   swap file or partition).
 - Extended runtime statistics which are easier to understand.
   Statistics will show memory percentages and the estimated savings.

Signed-off-by: Oever González <notengobattery@gmail.com>