kernel: add support for zram with writeback and zswap
Writeback for zram will allow zram to evict pages that are hard to
compress to a backing block device. This will reduce the memory used
both by the compressor (i.e. the overhead) and the data itself.

This commit also adds support for zswap, a compressed cache for swap
devices. The use of zswap can reduce the system latency under pressure
because it's known than it's most performant than zram.

Also, it will add the zsmalloc package as a separate package, because
it's used by both zram and zswap and they must be able to be used
separatedly. If zsmalloc is not an extra package, either they will have
to depend in each other or they must conflict.

CONFIG_PGTABLE_MAPPING may be a huge performance boost for some
architectures. It's known to be faster in ARM, therefore is enabled by
default when building for ARM. However, the user is able to disable the
feature from the menu or by modifying the configuratio.

The addition of zswap requieres the following extra modules to be
installed: z3fold, zbud and zpool. zram and zswap share the same
compressors, therefore there is no need to enable more packages.

Signed-off-by: Oever González <notengobattery@gmail.com>